+++ packages/next/src/server/app-render/has-loading-component-in-tree.tsx
@@ -1,0 +1,13 @@
+import type { LoaderTree } from '../lib/app-dir-module'
+
+export function hasLoadingComponentInTree(tree: LoaderTree): boolean {
+  const [, parallelRoutes, { loading }] = tree
+
+  if (loading) {
+    return true
+  }
+
+  return Object.values(parallelRoutes).some((parallelRoute) =>
+    hasLoadingComponentInTree(parallelRoute)
+  ) as boolean
+}
